requirements.txt
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0
Flask-Babel==0.11.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pbr==4.0.2
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.10.0
stevedore==1.24.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.14.1
